requirements.txt
asgiref==3.5.2
Django==4.1
gunicorn==20.1.0
python-dotenv==0.20.0
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
